+++ src/helpers/retry.ts
@@ -1,3 +1,6 @@
+import ms, { StringValue } from "ms";
+import OpenAI from "openai";
+
 interface RetryOptions {
   maxRetries: number;
   onError?: (error: unknown) => void | Promise<void>;
@@ -33,4 +36,22 @@ export async function retry<T>(fn: () => Promise<T>, options: RetryOptions): Pro
     delay *= 2;
   }
   throw lastError;
+}
+
+export function checkLlmRetryableState(error: unknown) {
+  if (error instanceof OpenAI.APIError && error.status) {
+    if ([500, 503].includes(error.status)) {
+      return true;
+    }
+    if (error.status === 429 && error.headers) {
+      const retryAfterTokens = error.headers["x-ratelimit-reset-tokens"];
+      const retryAfterRequests = error.headers["x-ratelimit-reset-requests"];
+      if (!retryAfterTokens || !retryAfterRequests) {
+        return true;
+      }
+      const retryAfter = Math.max(ms(retryAfterTokens as StringValue), ms(retryAfterRequests as StringValue));
+      return Number.isFinite(retryAfter) ? retryAfter : true;
+    }
+  }
+  return false;
 }
